scripts

+++ scripts/clean_encoders_npz.py
@@ -1,3 +1,4 @@
+import numpy as np
 from pathlib import Path
 from shutil import rmtree
 
@@ -8,17 +9,18 @@ speaker_dirs = [f for f in root_dir.glob("*") if f.is_dir()]
 # print("Speakers: {0}".format(speaker_dirs[0:10]))
 
 for speaker_dir in speaker_dirs:
-    mels = [f for f in speaker_dir.glob("*.npy") if f.is_file()]
+    mels = np.load(speaker_dir.joinpath("combined.npz"))
+    mels = mels.files
+
     # print(" - mels: {0}".format(len(mels)))
     remove_dir = False
     if len(mels) < 1:
         # print(" - NO MELS! {}".format(speaker_dir))
         remove_dir = True
-    elif len(mels) < 7:
+    elif len(mels) < 5:
         # print(" - Not enough MELS ({0}) {1}".format(len(mels), speaker_dir))
         remove_dir = True
 
-    if remove_dir == True:
+    if remove_dir == True and AUTO_REMOVE_DIR == True:
         print(" - Removing: {}".format(speaker_dir))
-        if AUTO_REMOVE_DIR == True:
-            rmtree(speaker_dir)
+        rmtree(speaker_dir)
